// frontend/src/app/core/models/user.model.ts
export class User {
    fullName?:       string;
    email?:          string;
    profilePicture?: null;
    specialization?: string;
    qualification?:  string;
    experience?:     string;
    headline?:       string;
    aboutMe?:        string;
    showProfile?:    boolean;
    showCourses?:    boolean;
    websiteUrl?:     null;
    facebookUrl?:    null;
    twitterUrl?:     null;
    linkedInUrl?:    null;
    instagramUrl?:   null;
    youtubeUrl?:     null;
}